Remove URL from bottle stanza
This causes us to download bottles straight from Homebrew's servers rather than forcing us to build our own and host them somewhere.  This is only possible when we don't need to apply any local julia-specific patches and can use the Homebrew bottles directly.  The only reason we need to have this file at all is because we need to override the `pour_bottle?` check within the normal Homebrew mainline formula.